refactor shutdown/restart methods to a new helper: AppLifetimeHelper.cs
