+++ WalletWasabi.Fluent/CrashReport/CrashReporter.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Avalonia;
 using Avalonia.Controls.ApplicationLifetimes;
+using WalletWasabi.Fluent.Helpers;
 using WalletWasabi.Logging;
 using WalletWasabi.Microservices;
 using WalletWasabi.Models;
@@ -20,34 +21,12 @@ namespace WalletWasabi.Fluent.CrashReport
 				var serializedException = exceptionToReport.ToSerializableException();
 				var base64ExceptionString = SerializableException.ToBase64String(serializedException);
 				var args = $"crashreport -exception=\"{base64ExceptionString}\"";
-				StartWasabiWithArgs(args);
+				AppLifetimeHelper.StartAppWithArgs(args);
 			}
 			catch (Exception ex)
 			{
 				Logger.LogWarning($"There was a problem while invoking crash report: '{ex}'.");
 			}
-		}
-
-		public static void RestartWasabi()
-		{
-			Task.Run(async () =>
-			{
-				await Task.Delay(1000);
-				ShutdownWasabi();
-			});
-
-			StartWasabiWithArgs();
-		}
-
-		public static void StartWasabiWithArgs(string args = "")
-		{
-			var path = Process.GetCurrentProcess().MainModule?.FileName;
-			if (string.IsNullOrEmpty(path))
-			{
-				throw new InvalidOperationException($"Invalid path: '{path}'");
-			}
-			var startInfo = ProcessStartInfoFactory.Make(path, args);
-			using var p = Process.Start(startInfo);
 		}
 
 		public static bool TryGetExceptionFromCliArgs(string[] args, [NotNullWhen(true)] out SerializableException? exception)
@@ -83,9 +62,5 @@ namespace WalletWasabi.Fluent.CrashReport
 			return false;
 		}
 
-		public static void ShutdownWasabi()
-		{
-			(Application.Current.ApplicationLifetime as IClassicDesktopStyleApplicationLifetime)?.Shutdown();
-		}
 	}
 }
